+++ src/main/java/net/fabricmc/mappingio/format/srg/SrgFileWriter.java
@@ -63,6 +63,10 @@ public final class SrgFileWriter implements MappingWriter {
 
 	@Override
 	public boolean visitField(String srcName, @Nullable String srcDesc) throws IOException {
+		if (xsrg && srcDesc == null) {
+			return false;
+		}
+
 		memberSrcName = srcName;
 		memberSrcDesc = srcDesc;
 		memberDstName = null;
@@ -73,6 +77,10 @@ public final class SrgFileWriter implements MappingWriter {
 
 	@Override
 	public boolean visitMethod(String srcName, @Nullable String srcDesc) throws IOException {
+		if (srcDesc == null) {
+			return false;
+		}
+
 		memberSrcName = srcName;
 		memberSrcDesc = srcDesc;
 		memberDstName = null;
@@ -123,11 +131,11 @@ public final class SrgFileWriter implements MappingWriter {
 			write("CL: ");
 			break;
 		case FIELD:
-			if (memberSrcDesc == null || memberDstName == null || (xsrg && memberDstDesc == null)) return false;
+			if (memberDstName == null || xsrg && memberDstDesc == null) return false;
 			write("FD: ");
 			break;
 		case METHOD:
-			if (memberSrcDesc == null || memberDstName == null || memberDstDesc == null) return false;
+			if (memberDstName == null || memberDstDesc == null) return false;
 			write("MD: ");
 			break;
 		default: