+++ src/test/java/net/fabricmc/mappingio/test/TestMappings.java
@@ -55,7 +55,9 @@ public class TestMappings {
 		if (delegate.visitHeader()) {
 			delegate.visitNamespaces(MappingUtil.NS_SOURCE_FALLBACK, Arrays.asList(MappingUtil.NS_TARGET_FALLBACK, MappingUtil.NS_TARGET_FALLBACK + "2"));
 			delegate.visitMetadata("name", "valid");
-			delegate.visitMetadata(MigrationMapConstants.ORDER_KEY, "0");
+			delegate.visitMetadata(MigrationMapConstants.ORDER_KEY, MigrationMapConstants.DEFAULT_ORDER);
+			delegate.visitMetadata("property-with-empty-value", "");
+			delegate.visitMetadata("property-with-null-value", null);
 		}
 
 		if (delegate.visitContent()) {