+++ CHANGELOG.md
@@ -4,10 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added a simplified `MappingNsCompleter` constructor for completing all destination names with the source names
 - Made `OuterClassNamePropagator` configurable
 - Made Enigma writer always output destination names if visited explicitly, establishing consistency across all writers
-- Added a simplified `MappingNsCompleter` constructor for completing all destination names with the source names
 - Adjusted format detection to only return ENIGMA_DIR for non-empty directories with at least one `.mapping` file
+- Fixed writer NPEs when metadata or member source descriptors are null
+- Fixed SRG writer omitting fields with missing source descriptors
 
 ## [0.7.1] - 2025-01-07
 - Restored the ability to read source-namespace-only mapping files, even if not spec-compliant